+++ src/App.js
@@ -1,4 +1,4 @@
-import LottoGame from './LottoGame.js';
+import LottoGame from './Controller/LottoGame.js';
 
 class App {
   #lottoGame = new LottoGame();
