+++ src/View/View.js
@@ -1,7 +1,6 @@
 import InputView from './InputView.js';
 import OutputView from './OutputView.js';
 import MESSAGE from '../constants/Message.js';
-import { Validator } from '../Validator.js';
 
 class View {
   #inputView = InputView;
@@ -9,10 +8,7 @@ class View {
   #outputView = OutputView;
 
   async readPurchaseAmount() {
-    const userInput = await this.#inputView.inputLine(MESSAGE.PURCHASE_AMOUNT);
-    Validator.validatePurchaseAmount(userInput);
-
-    return userInput;
+    return await this.#inputView.inputLine(MESSAGE.PURCHASE_AMOUNT);
   }
 
   async readWinningNumber() {
